Add files via upload
As you can see from the title this is one of the final builds of the program. There are previous builds but i'm not going to upload them since they were half finished. In this new version I have added the ability to mark reminders as done, as well as the ability to edit and view reminders once they have been made. Another build will be on the way, in that build i'm just gonna done some testing and fix up any bugs.